Move inet stuff in its own file
